+++ index.js
@@ -34,6 +34,12 @@ var REGEXP_CACHE = Object.create(null)
 var REGEXP_ESCAPE_CHARS_REGEXP = /[\^$\\.*+?()[\]{}|]/g
 
 /**
+ * RegExp to match basic restricted characters for loose validation.
+ */
+
+var RESTRICTED_CHARS_REGEXP = /[;=]/
+
+/**
  * RegExp to match Same-Site cookie attribute value.
  */
 
@@ -131,11 +137,11 @@ Cookies.prototype.set = function(name, value, opts) {
 };
 
 function Cookie(name, value, attrs) {
-  if (!fieldContentRegExp.test(name)) {
+  if (!fieldContentRegExp.test(name) || RESTRICTED_CHARS_REGEXP.test(name)) {
     throw new TypeError('argument name is invalid');
   }
 
-  if (value && !fieldContentRegExp.test(value)) {
+  if (value && (!fieldContentRegExp.test(value) || RESTRICTED_CHARS_REGEXP.test(value))) {
     throw new TypeError('argument value is invalid');
   }